ai/requirements.txt
python-dotenv
langchain
langchain_community
langchain_huggingface
sentence-transformers
langchain_chroma
chromadb
gigachat
langchain-gigachat
gigachain-cli
